ci-github: Handle forced pushes and branch deletions
Cancel CI of the previous head commit in the case of overwritten or deleted
history.

In the process, move branch push handling from handle_check_suite_request() to
handle_branch_push() because GitHub sends a push event instead of a
check_suite when a new branch is created so tenant reference counting would
not work otherwise.