Automatically increase build queued timeout for big number of builds in 'queued' notifications
Note that this commit changes the semantics of the tenant::queued_timestamp
member. Previously it was set to the time when the build objects are turned
into the queued state. Now it is set to the time when the queued notification
for them is assumed to be delivered to the third party service.